Download and unzip in parallel.
This change uses HTTP range requests to allow downloading a zip file to occur
in parallel with actually unzipping it.  A zip file has its directory at the
end, then content earlier in the file. If an HTTP server advertises that it
supports the Range header, then we can use that to first read the directory and
thus kick off threads to unzip each file, reading the data at the same time.