Comparitive fuzzing of ripunzip and zip-rs.
This commit adds a comparitive fuzzer designed to ensure that zip-rs and
ripunzip produce the same outcome each time. The goal isn't to explore the
fundamental unzipping algorithm, but to look for cases in which the parallel
unzipping results in problems.

A subsequent commit will extend this to the HTTP download too.