Various fixes discovered through comparitive fuzzing
A future set of PRs will add comparitive fuzzing infrastructure
to compare the behavior of ripunzip with the underlying zip-rs
behavior.

This commit includes a handful of fixes discovered that way.
Some of them are legitimate bug fixes; others move our behavior
closer to that of zip-rs in order that fuzzing doesn't give us
false positives.