Add comparative benchmark.
This commit adds a 'cargo criterion' benchmark suite for ripunzip.
As this shares a fair bit of code with the fuzzer and the basic unit test
suite, it restructures the project to create a test_utils workspace member.
(Unlike fuzzing, there's no way to detect use of a bench by means of
cfg directives, so no way to optionally enable this code in the main
project... in any case this is probably cleaner.)

This also adds tests for each of the three server types we use.